# src/main/resources/static/openapi/banking-openapi.yaml
openapi: 3.0.3
info:
  title: Brite Banking API
  version: 1.0.0
  description: |
    REST API for the `org.brite.banking` module of the bankingservices Spring Boot application:
    account registration, lookup, deposit/withdrawal, closing (single and bulk), paginated
    account-status search, bank office/ATM location search, bank statements, plus the module's async-notification and
    Resilience4j payment demo endpoints.

    ## Gateway behavior (applies to every endpoint in this document)

    Every path below sits behind two servlet filters (`org.brite.banking.gateway`), which run
    *before* the controllers:

    1. **`BusinessTransactionIdFilter`** generates a UUID per request and returns it in the
       `X-BTID` response header. The same id is stamped on every server log line for that
       request, so quote it when reporting a problem.
    2. **`BankingRateLimitFilter`** requires the customer-id header (default `X-Customer-Id`,
       configurable via `banking.rate-limit.customer-header-name`) and caps each customer at
       `banking.rate-limit.requests-per-day` requests/day (default 1000). A missing/blank
       header is rejected with `400`; an exhausted quota with `429`. Both are **plain-text**
       bodies written by the filter itself, not the JSON/text bodies produced by the
       controllers' exception handler. Every non-rejected response carries
       `X-RateLimit-Limit` and `X-RateLimit-Remaining`.

...
    authentication - the API has no auth layer.

...

...

...

...
servers:
...
security:
  - CustomerId: []
paths:
...
  /v1/api/accounts/withdraw:
    post:
      tags: [Accounts]
      operationId: withdraw
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Withdraw funds
      description: |
        Withdraws `withdrawAmount` from a checking or savings account. The account type is
        derived from the account-number prefix (`CH` -> `CHECKING`, `SV` -> `SAVINGS`) and
        `accountType` in the body **must match it**.

        Checks, in order: account exists; account is not `CLOSED`; sufficient funds; the
        remaining balance must stay at or above the minimum
        (`banking.constraints.checking-minimum-balance` 25.00 / `saving-minimum-balance` 100.00
        by default). A completed withdrawal is recorded in withdrawal history and the
        transaction ledger, and triggers an async email/SMS.

        `withdrawalDate` and `withdrawalStatus` are accepted but ignored - the server stamps
        today's date and `COMPLETED`.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WithdrawalRequest'}
            example:
              accountNumber: CH-0000088291
              accountType: CHECKING
              withdrawAmount: 100.75
              firstName: Jane
              lastName: Doe
              street: '111'
              addressLine1: Leafvillage
              city: Leander
              state: TX
              zip: '78717'
      responses:
        '200':
          description: Withdrawal completed; returns the account with its updated balance.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: >-
            Validation failure (JSON), or a business rule failed (plain text): non-positive
            amount, unrecognized prefix, type/prefix mismatch, account closed or suspended,
            insufficient funds, or minimum balance would be violated.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
              examples:
                insufficientFunds:
                  value: Insufficient funds in account CH-0000088291
                minimumBalance:
                  value: 'Withdrawal declined: account CH-0000088291 must retain a minimum balance of 25.00'
                closed:
                  value: Account CH-0000010004 is closed and cannot be used for transactions
                suspended:
                  value: Account CH-0000050001 is suspended and cannot be used for transactions until it is reactivated
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/deposit:
    post:
      tags: [Accounts]
      operationId: deposit
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Deposit funds
      description: |
        Deposits `amount` into a checking or savings account. As with withdrawals, the
        account type comes from the account-number prefix and `accountType` must match it.

        `depositType` must be `cash` or `check` (case-insensitive). **Cash deposits are
        capped** at `banking.constraints.maximum-deposit-amount-by-cash` (default 5000.00);
        check deposits are not. A `CLOSED` or `SUSPENDED` account rejects deposits (`400`). Unlike withdrawals,
        deposits do not trigger a notification.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DepositForm'}
            example:
              accountNumber: SV-0000044102
              accountType: SAVINGS
              amount: 250.00
              depositType: check
              firstName: Jane
              lastName: Doe
              street: '111'
              addressLine1: Leafvillage
              city: Leander
              state: TX
              zip: '78717'
      responses:
        '200':
          description: Deposit completed; returns the account with its updated balance.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: >-
            Validation failure (JSON), or a business rule failed (plain text): non-positive
            amount, unrecognized prefix, type/prefix mismatch, invalid deposit type, cash
            limit exceeded, or account closed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
              examples:
                cashLimit:
                  value: Cash deposits cannot exceed 5000.00
                depositType:
                  value: Deposit type must be 'cash' or 'check'
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/{accountNumber}/close:
    post:
      tags: [Accounts]
      operationId: closeAccount
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Close one account
      description: >-
        Transitions an `ACTIVE` or `SUSPENDED` account to `CLOSED` and stamps `closedDate` with
        today (a suspension is cleared). There is no reopen operation. Evicts the account-search cache.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: 'Account closed; returns it with `accountStatus: CLOSED`.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: Account is already closed. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
              example: Account CH-0000010004 is already closed
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/{accountNumber}/suspend:
    post:
      tags: [Account suspension]
      operationId: suspendAccount
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Suspend an account
      description: |
        `ACTIVE` -> `SUSPENDED`. While suspended the account rejects **every** withdraw and
        deposit (`400`) until it is reactivated, or until `endDateTime` passes and the expiry job
        (`banking.suspension.expiry-job`, every 60s) reactivates it. `endDateTime` omitted = indefinite.

        `startDateTime` defaults to now and must not be in the future (backdating is allowed).
        `endDateTime`, if given, must be after the start and in the future. Both are ISO local
        date-times (`2026-10-01T09:30:00`). Evicts the account-search cache.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SuspendAccountRequest'}
            example:
              notes: Suspected fraudulent card activity - under review
              endDateTime: '2026-12-31T17:00:00'
      responses:
        '200':
          description: 'Account suspended; returned with `accountStatus: SUSPENDED` and the suspension fields.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: >-
            Validation failure (JSON, e.g. blank notes), or a business rule failed (plain text):
            account closed, already suspended, start in the future, end not after start or in the past.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
              examples:
                alreadySuspended:
                  value: Account CH-0000050001 is already suspended
                closed:
                  value: Account CH-0000010004 is closed and cannot be suspended
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/{accountNumber}/suspension:
    patch:
      tags: [Account suspension]
      operationId: updateSuspension
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Update a current suspension
      description: |
        Changes the end and/or notes of a suspension; only the fields supplied change, and at
        least one is required. The new `endDateTime` must be in the future and after the stored
        start. The suspension's start can't be changed, and an end can't be removed once set
        (reactivate instead). Evicts the account-search cache.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateSuspensionRequest'}
            example:
              endDateTime: '2027-01-31T17:00:00'
              notes: Extended pending compliance review
      responses:
        '200':
          description: Updated account.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: >-
            Validation failure (JSON), or a business rule failed (plain text): nothing to update,
            account isn't suspended, new end in the past or not after the start.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SpringErrorResponse'}
            text/plain:
              schema: {type: string}
              example: Account CH-0000010001 is not suspended
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/accounts/{accountNumber}/reactivate:
    post:
      tags: [Account suspension]
      operationId: reactivateAccount
      security:
        - CustomerId: []
          CustomerBearerAuth: []
      summary: Reactivate a suspended account
      description: |
        `SUSPENDED` -> `ACTIVE`; the suspension flag, dates and notes are cleared and the
        account can transact again. Evicts the account-search cache.
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: Account is `ACTIVE` again.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: Account isn't suspended. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
              example: Account CH-0000010001 is not suspended
        '401':
          $ref: '#/components/responses/CustomerUnauthorized'
        '403':
          $ref: '#/components/responses/CustomerForbidden'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /v1/api/staff/accounts/withdraw:
    post:
      tags: [Staff]
      operationId: staffWithdraw
      summary: Withdraw as an employee
      description: |
        Requires the `WITHDRAW` privilege. On success it delegates to the same rule set as the matching customer endpoint, so its business errors apply too.
      security:
        - CustomerId: []
          BearerAuth: []
      parameters:
        - name: locationId
          in: query
          required: false
          description: Branch office or ATM where this is handled; defaults to the employee's own branch (none for an area manager). `404` if unknown. The employee and this location are recorded on the transaction.
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WithdrawalRequest'}
      responses:
        '200':
          description: The updated account.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: A failed validation (JSON) or a business rule (plain text).
          content:
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmployeeForbidden'
        '404':
          description: The token's employee no longer exists, or unknown account. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/staff/accounts/deposit:
    post:
      tags: [Staff]
      operationId: staffDeposit
      summary: Deposit as an employee
      description: |
        Requires the `DEPOSIT` privilege. On success it delegates to the same rule set as the matching customer endpoint, so its business errors apply too.
      security:
        - CustomerId: []
          BearerAuth: []
      parameters:
        - name: locationId
          in: query
          required: false
          description: Branch office or ATM where this is handled; defaults to the employee's own branch (none for an area manager). `404` if unknown. The employee and this location are recorded on the transaction.
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DepositForm'}
      responses:
        '200':
          description: The updated account.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: A failed validation (JSON) or a business rule (plain text).
          content:
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmployeeForbidden'
        '404':
          description: The token's employee no longer exists, or unknown account. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/staff/accounts/{accountNumber}/suspend:
    post:
      tags: [Staff]
      operationId: staffSuspendAccount
      summary: Suspend an account as an employee
      description: |
        Requires the `SUSPEND_ACCOUNT` privilege. On success it delegates to the same rule set as the matching customer endpoint, so its business errors apply too.
      security:
        - CustomerId: []
          BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SuspendAccountRequest'}
      responses:
        '200':
          description: The updated account.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: A failed validation (JSON) or a business rule (plain text).
          content:
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmployeeForbidden'
        '404':
          description: The token's employee no longer exists, or unknown account. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/staff/accounts/{accountNumber}/suspension:
    patch:
      tags: [Staff]
      operationId: staffUpdateSuspension
      summary: Update a suspension as an employee
      description: |
        Requires the `UPDATE_SUSPENSION` privilege. On success it delegates to the same rule set as the matching customer endpoint, so its business errors apply too.
      security:
        - CustomerId: []
          BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateSuspensionRequest'}
      responses:
        '200':
          description: The updated account.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: A failed validation (JSON) or a business rule (plain text).
          content:
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmployeeForbidden'
        '404':
          description: The token's employee no longer exists, or unknown account. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/staff/accounts/{accountNumber}/reactivate:
    post:
      tags: [Staff]
      operationId: staffReactivateAccount
      summary: Reactivate an account as an employee
      description: |
        Requires the `REACTIVATE_ACCOUNT` privilege. On success it delegates to the same rule set as the matching customer endpoint, so its business errors apply too.
      security:
        - CustomerId: []
          BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: The updated account.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: A failed validation (JSON) or a business rule (plain text).
          content:
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmployeeForbidden'
        '404':
          description: The token's employee no longer exists, or unknown account. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/staff/accounts/{accountNumber}/close:
    post:
      tags: [Staff]
      operationId: staffCloseAccount
      summary: Close an account as an employee
      description: |
        Requires the `CLOSE_ACCOUNT` privilege. On success it delegates to the same rule set as the matching customer endpoint, so its business errors apply too.
      security:
        - CustomerId: []
          BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: The updated account.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        '400':
          description: A failed validation (JSON) or a business rule (plain text).
          content:
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmployeeForbidden'
        '404':
          description: The token's employee no longer exists, or unknown account. Plain-text body.
          content:
            text/plain:
              schema: {type: string}
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/staff/employees:
    get:
      tags: [Staff]
      operationId: listEmployees
      summary: List employees
      description: |
        Requires `MANAGE_EMPLOYEES` (area managers). Optional `role` filter; sortable by `lastName` (default), `firstName`, `employeeNumber`, `role`, `hireDate`; default `size=20`.
      security:
        - CustomerId: []
          BearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema: {$ref: '#/components/schemas/EmployeeRole'}
        - {name: page, in: query, required: false, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, required: false, schema: {type: integer, minimum: 1, default: 20}}
        - {name: sort, in: query, required: false, schema: {type: string}, example: 'lastName,asc'}
      responses:
        '200':
          description: A Spring Data page of employee profiles (`content` holds `Employee` objects).
          content:
            application/json:
              schema: {type: object}
        '400':
          description: An unsupported `sort` property. Plain text.
          content:
            text/plain:
              schema: {type: string}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmployeeForbidden'
        '404':
          description: The token's employee no longer exists. Plain text.
          content:
            text/plain:
              schema: {type: string}
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/staff/employees/{employeeNumber}:
    get:
      tags: [Staff]
      operationId: getEmployee
      summary: Get one employee profile
      description: |
        An employee can read their own profile; reading anyone else's requires `MANAGE_EMPLOYEES`.
      security:
        - CustomerId: []
          BearerAuth: []
      parameters:
        - name: employeeNumber
          in: path
          required: true
          schema: {type: string}
          example: EMP-000010
      responses:
        '200':
          description: The employee profile, including the privileges derived from the role.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Employee'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmployeeForbidden'
        '404':
          description: The token's employee or the requested employee not found. Plain text.
          content:
            text/plain:
              schema: {type: string}
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    CustomerId:
      type: apiKey
      in: header
      name: X-Customer-Id
      description: |
        Customer identifier used solely for per-customer daily rate limiting - **not**
        authentication; any non-blank value is accepted. The header name is configurable
        (`banking.rate-limit.customer-header-name`) and the check can be turned off with
        `banking.rate-limit.enabled: false`.

    CustomerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The customer access token returned by `POST /v1/api/customers/login` (HS256, 30 minutes, no refresh), sent as
        `Authorization: Bearer <token>`. Required on every `/v1/api/accounts/**` and `/bff/v1/portal/**` operation except
        the two that create a customer (`POST /v1/api/accounts/newaccount`, `POST /bff/v1/portal/accounts/open`). It proves who
        logged in; the customer's login is re-checked on every request, and the customer can only reach their own accounts.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The employee access token returned by `POST /v1/api/staff/login` (HS256, 30 minutes, no refresh), sent as
        `Authorization: Bearer <token>`. It proves who logged in; permissions and status are re-checked from the
        database on every request, so a suspended login stops working immediately.
...
  responses:
...
    AccountNotFound:
      description: No account with that number. Plain-text body.
      content:
        text/plain:
          schema: {type: string}
          example: 'Account not found: CH-9999999999'
...